Fix Earth-scale terrain visibility and add World Creator-style per-layer masking
The "terrain invisible at Earth scale" report traced to mesh resolution, not
height math: Max Subdivisions is a hard cap independent of Radius, so at
Earth's radius default settings gave ~104km triangle edges against
meter-scale terrain. Fixes float->double radius precision throughout the
icosphere generator (float was losing tens of centimeters at Earth scale),
widens the Max Subdivisions ceiling with clearer diagnostics, and documents
that ground-level detail at planetary radius needs a future chunked/LOD
system rather than a bigger single mesh.

Also lets any layer in the main Terrain Layer Stack carry its own procedural
mask (reusing the existing Climate/Composite Mask types), so filters like
Erosion or Noise can be region-restricted directly - closer to World
Creator's per-filter masking model - instead of only through the separate
whole-biome TerrainDetail stack. Fixes a companion bug where a stack's first
enabled layer used its own blend mode against an uninitialized accumulator
(Multiply always yielded 0, Min/Max clipped/floored to 0).

Also fixes several bugs found during the review: double biome-mask
evaluation per vertex in the biome pass, no Min<=Max validation on mask
ranges, redundant per-vertex noise seed-hash recomputation, and TerrainDetail
stacks never receiving Sea Level/Radius context. Centralizes five
independently-duplicated equirectangular grid-walk/bilinear-sample
implementations into one shared FSolarOrbzLatLongGrid utility.

+++ Source/SolarOrbz/Private/SolarOrbzEditor.cpp
@@ -152,13 +152,13 @@ void SSolarOrbzMainPanel::Construct(const FArguments& InArgs)
 			+ SVerticalBox::Slot().AutoHeight().Padding(0.0f, 2.0f)
 			[
 				MakeLabeledRow(LOCTEXT("RadiusLabel", "Radius (m)"),
-					SNew(SSpinBox<float>)
-					.MinValue(0.01f)
-					.MinSliderValue(1.0f)
-					.MaxSliderValue(1000000.0f) // slider convenience range (1,000 km) - typing goes further, no hard ceiling
-					.Delta(1.0f)
+					SNew(SSpinBox<double>)
+					.MinValue(0.01)
+					.MinSliderValue(1.0)
+					.MaxSliderValue(1000000.0) // slider convenience range (1,000 km) - typing goes further, no hard ceiling
+					.Delta(1.0)
 					.Value_Lambda([this]() { return RadiusMeters; })
-					.OnValueChanged_Lambda([this](float NewValue) { RadiusMeters = NewValue; })
+					.OnValueChanged_Lambda([this](double NewValue) { RadiusMeters = NewValue; })
 				)
 			]
