+++ Source/SolarOrbz/Public/SolarOrbzEditor.h
@@ -78,7 +78,7 @@ public:
 
 private:
 	// --- Staged parameters, pushed onto the preview actor when Generate is pressed. ---
-	float RadiusMeters = 1000.0f;
+	double RadiusMeters = 1000.0;
 	float VerticesPerMeter = 1.0f;
 	int32 MaxSubdivisions = 6;
 	bool bEnablePreviewCollision = false;
